Backend/requirements-docker.txt
# Core
Django>=4.2,<6
djangorestframework>=3.14,<4
django-cors-headers>=4.3,<5
django-environ>=0.11,<1
dj-database-url>=2.1,<3
psycopg2-binary>=2.9,<3
asgiref>=3.7,<4

# ASGI / Channels
daphne>=4.0,<5
channels>=4.0,<5
channels-redis>=4.2,<5

# Celery
celery[redis]>=5.3,<6
redis>=5.0,<6
kombu>=5.3,<6
billiard>=4.2,<5
vine>=5.1,<6

# SNMP (device discovery/polling)
# PySNMP v7+ uses v3arch asyncio API for high-performance concurrent polling
# Supports SNMP v1/v2c/v3 protocols
# Required OIDs: RFC 3805 (Printer MIB v2), RFC 2790 (Host Resources MIB)
pysnmp>=7.1,<8

# Production server
gunicorn>=21,<23
whitenoise>=6.5,<7

# Optional: keep if your code uses them
requests>=2.31,<3
python-dotenv>=1.0,<2
